WIP: trying to everybody running with new model
New model uses a plugin for configuration and extensions for an
publishing implementation.

Included invoker plugin to some high level tests.

Risks:
Getting the "current project" from the EventSession could cause
problems since I'm getting the "top level project"...don't really know
waht that means.

Manually parsing the configuratino of the plugin. I'd like to convert
the xml int the Mojo and have it loaded for me. Not sure how to do that.